Move android mod to seperate file
